% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conox_gkm.R
\name{LDV_g_km}
\alias{LDV_g_km}
\title{Calculate g/km emissions for light duty vehicles}
\usage{
LDV_g_km(rs_data = conox)
}
\arguments{
\item{rs_data}{A data frame input.}
}
\value{

}
\description{
Calculate g/km emissions for light duty vehicles
}
